Swift 6 preparedness (#54)
Motivation

We should feel confident that we compile clean under the strict
checks of Swift 6. As this is a data-processing-only framework
it's not a surprise that it turns out to be fine, but this patch
adds the necessary docker-compose steps to ensure it continues
to work.

Modifications

- Add docker-compose file for Swift 6 nightlies
- Extend docker-compose to pass explicit sendable and strict
    concurrency checks
- Add those checks to 6.0 and main nightlies

Result

Better confidence that we don't break our downstreams in Swift 6
support.