Update Swift targets to enable ExistentialAny
Motivation:

With the advent of Upcoming Feature Flags https://www.swift.org/blog/using-upcoming-feature-flags/ ,
we should make it a one-liner to enable feature flags.

As a follow-on to PR 63, we should enable the ExistentialAny feature flag, as
the code currently complies, and we may as well help flag any unintentional
regressions to that behavior.

Modifications:

Update Package.swift to allow for one-liner additions of new feature flags.
Enable the upcoming feature flag ExistentialAny

Result:

Better confidence that we don't break downstreams that compile this code with
ExistentialAny enabled.
Makes it (slightly) easier for future authors to modify the set of enabled feature flags.