Migrate to GitHub Actions reusable workflows (1)
Motivation:

Migrate to GitHub Actions reusable workflows whilst keeping CI green.

Modifications:

As the first step in migrating to use GitHub actions this PR provides
the workflows which will be used in the future, at first in disabled
state so that we can keep CI green during the changeover.

Result:

Once this PR is merged we can:
* Enable the new CI
* In a future PR neutralize the old CI scripts
* Disable the old CI via GitHub
* In a future PR Delete the old CI scripts